% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/another_product.R
\name{another_product}
\alias{another_product}
\title{Another product}
\usage{
another_product(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
Cross product of x and y
}
\description{
Another product
}
\examples{
crossprod(c(1,2,3),c(4,5,6))
}
