% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\docType{data}
\name{m}
\alias{m}
\title{A matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 5 columns.
}
\usage{
m
}
\description{
A matrix provided as data in tp2 package
}
\keyword{datasets}
