% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_by_element.R
\name{product_by_element}
\alias{product_by_element}
\title{Element-wise product}
\usage{
product_by_element(x, y)
}
\arguments{
\item{x}{scalar, vector or matrix}

\item{y}{scalar, vector or matrix}
}
\value{
The element-wise product of \code(x) and \code(y)
}
\description{
Element-wise product
}
\examples{
product_by_element(c(1,2,3),c(4,5,6))
}
\author{
Alexey Larionov
}
