% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_product.R
\name{dot_product}
\alias{dot_product}
\title{Dot product}
\usage{
dot_product(x, y)
}
\arguments{
\item{x}{scalar, vector or matrix}

\item{y}{scalar, vector or matrix}
}
\value{
The dot product of \code(x) and \code(y)
}
\description{
Dot product
}
\examples{
product_by_element(c(1,2,3),matrix(c(1:12), nrow=3)
}
\author{
Alexey Larionov
}
