% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_by_element.R
\name{product_by_element}
\alias{product_by_element}
\title{Product by element}
\usage{
product_by_element(x, y)
}
\arguments{
\item{x}{scalar, vector or matrix}

\item{y}{scalar, vector or matrix}
}
\value{
The element-wise product of x and y
}
\description{
Product by element
}
\author{
Alexey Larionov
}
